Changed integers to bigint

+++ models/UsersShop.js
@@ -1,9 +1,9 @@
 module.exports = (sequelize, DataTypes) => {
   return sequelize.define('users_shop', {
     userId: DataTypes.STRING,
-    itemId: DataTypes.INTEGER,
+    itemId: DataTypes.BIGINT,
     amount: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.BIGINT,
       allowNull: false,
       'default': 0,
     },
